+++ node/build.rs
@@ -1,4 +1,4 @@
-use substrate_build_script_utils::{generate_cargo_keys, rerun_if_git_head_changed};
+use polkadot_sdk::substrate_build_script_utils::{generate_cargo_keys, rerun_if_git_head_changed};
 
 fn main() {
 	generate_cargo_keys();
