+++ pallets/hybrid-orderbook/src/mock.rs
@@ -15,30 +15,35 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Test environment for Asset Conversion pallet.
-
 use super::*;
 use crate as pallet_hybrid_orderbook;
+
 use core::default::Default;
-use frame_support::{
-	construct_runtime, derive_impl,
-	instances::{Instance1, Instance2},
-	ord_parameter_types, parameter_types,
-	traits::{
-		tokens::{
-			fungible::{NativeFromLeft, NativeOrWithId, UnionOf},
-			imbalance::ResolveAssetTo,
+pub use polkadot_sdk::{
+	frame_support::{
+		assert_noop, assert_ok, construct_runtime, derive_impl,
+		instances::{Instance1, Instance2},
+		ord_parameter_types, parameter_types,
+		traits::{
+			fungible::Inspect as FungibleInspect,
+			fungibles::{Inspect, InspectEnumerable},
+			tokens::{
+				fungible::{NativeFromLeft, NativeOrWithId, UnionOf},
+				imbalance::ResolveAssetTo,
+			},
+			AsEnsureOriginWithArg, ConstU32, Get,
 		},
-		AsEnsureOriginWithArg, ConstU128, ConstU32,
+		PalletId,
 	},
-	PalletId,
-};
-use frame_system::{EnsureSigned, EnsureSignedBy};
-use sp_arithmetic::Permill;
-use sp_core::ConstU64;
-use sp_runtime::{
-	traits::{AccountIdConversion, BlockNumber, IdentityLookup},
-	BuildStorage,
+	frame_system::{EnsureSigned, EnsureSignedBy},
+	pallet_assets, pallet_balances,
+	sp_arithmetic::Permill,
+	sp_core::ConstU64,
+	sp_io::TestExternalities,
+	sp_runtime::{
+		traits::{AccountIdConversion, IdentityLookup},
+		BuildStorage,
+	},
 };
 
 type Block = frame_system::mocking::MockBlock<Test>;
@@ -174,7 +179,7 @@ impl Config for Test {
 	type BenchmarkHelper = ();
 }
 
-pub(crate) fn new_test_ext() -> sp_io::TestExternalities {
+pub(crate) fn new_test_ext() -> TestExternalities {
 	let mut t = frame_system::GenesisConfig::<Test>::default().build_storage().unwrap();
 
 	pallet_balances::GenesisConfig::<Test> {
@@ -183,7 +188,7 @@ pub(crate) fn new_test_ext() -> sp_io::TestExternalities {
 	.assimilate_storage(&mut t)
 	.unwrap();
 
-	let mut ext = sp_io::TestExternalities::new(t);
+	let mut ext = TestExternalities::new(t);
 	ext.execute_with(|| System::set_block_number(1));
 	ext
 }
