+++ runtime/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "warpx-runtime"
-version = "4.0.0-dev"
-description = "A fresh FRAME-based Substrate node"
+version = "0.1.0"
+description = "warp(x) = x`"
 authors.workspace = true
 homepage.workspace = true
 edition.workspace = true
@@ -12,129 +12,58 @@ publish = false
 targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
+smallvec = { workspace = true, default-features = true }
+codec = { workspace = true, features = ["derive"] }
+scale-info = { workspace = true, features = ["derive"] }
 
-codec = { package = "parity-scale-codec", version = "3.6.1", default-features = false, features = [
-    "derive",
+polkadot-sdk = { workspace = true, features = [
+    "experimental",
+    "pallet-assets",
+    "pallet-aura",
+    "pallet-balances",
+    "pallet-grandpa",
+    "pallet-sudo",
+    "pallet-timestamp",
+    "pallet-transaction-payment",
+    "pallet-transaction-payment-rpc-runtime-api",
+
+    "pallet-xcm",
+    "polkadot-parachain-primitives",
+    "polkadot-runtime-common",
+    "staging-xcm",
+    "staging-xcm-builder",
+    "staging-xcm-executor",
+
+    "cumulus-pallet-aura-ext",
+    "cumulus-pallet-session-benchmarking",
+    "cumulus-pallet-xcm",
+    "cumulus-pallet-xcmp-queue",
+    "cumulus-primitives-aura",
+    "cumulus-primitives-core",
+    "cumulus-primitives-storage-weight-reclaim",
+    "cumulus-primitives-utility",
+    "pallet-collator-selection",
+    "parachains-common",
+    "staging-parachain-info",
+
+    "runtime",
 ] }
-scale-info = { version = "2.10.0", default-features = false, features = [
-    "derive",
-] }
-
-frame-support = { workspace = true, features = ["experimental", "tuples-96"] }
-frame-system = { workspace = true }
-frame-try-runtime = { workspace = true, optional = true }
-frame-executive = { workspace = true }
-
-sp-api = { workspace = true }
-sp-block-builder = { workspace = true }
-sp-consensus-aura = { workspace = true }
-sp-consensus-grandpa = { workspace = true }
-sp-genesis-builder = { workspace = true }
-sp-core = { workspace = true }
-sp-inherents = { workspace = true }
-sp-offchain = { workspace = true }
-sp-runtime = { workspace = true }
-sp-session = { workspace = true }
-sp-std = { workspace = true }
-sp-transaction-pool = { workspace = true }
-sp-version = { workspace = true }
-
-pallet-assets = { workspace = true }
-pallet-aura = { workspace = true }
-pallet-balances = { workspace = true }
-pallet-grandpa = { workspace = true }
-pallet-indices = { workspace = true }
-pallet-example-tasks = { workspace = true }
-pallet-sudo = { workspace = true }
-pallet-timestamp = { workspace = true }
-pallet-transaction-payment = { workspace = true }
-
-# RPC
-frame-system-rpc-runtime-api = { workspace = true }
-pallet-transaction-payment-rpc-runtime-api = { workspace = true }
-
-# Benchmarking
-frame-benchmarking = { workspace = true, optional = true }
-frame-system-benchmarking = { workspace = true, optional = true }
 
 # Local
-pallet-hybrid-orderbook = { version = "4.0.0-dev", path = "../pallets/hybrid-orderbook", default-features = false }
-
-# Polkadot
-xcm = { workspace = true }
+pallet-hybrid-orderbook = { workspace = true }
 
 [build-dependencies]
-substrate-wasm-builder = { version = "17.0.0", git = "https://github.com/paritytech/polkadot-sdk.git", tag = "polkadot-v1.10.0", optional = true }
+polkadot-sdk = { optional = true, workspace = true, features = [
+    "substrate-wasm-builder",
+] }
 
 [features]
 default = ["std"]
 std = [
     "codec/std",
     "scale-info/std",
-    "substrate-wasm-builder",
-    "frame-try-runtime?/std",
-    "frame-system-benchmarking?/std",
-    "frame-benchmarking?/std",
-    "frame-executive/std",
-    "frame-support/std",
-    "frame-system-rpc-runtime-api/std",
-    "frame-system/std",
-    "frame-try-runtime/std",
-    "sp-api/std",
-    "sp-block-builder/std",
-    "sp-consensus-aura/std",
-    "sp-consensus-grandpa/std",
-    "sp-genesis-builder/std",
-    "sp-core/std",
-    "sp-inherents/std",
-    "sp-offchain/std",
-    "sp-runtime/std",
-    "sp-session/std",
-    "sp-std/std",
-    "sp-transaction-pool/std",
-    "sp-version/std",
-    "pallet-assets/std",
-    "pallet-aura/std",
-    "pallet-balances/std",
-    "pallet-grandpa/std",
-    "pallet-indices/std",
-    "pallet-example-tasks/std",
-    "pallet-sudo/std",
+    "polkadot-sdk/std",
     "pallet-hybrid-orderbook/std",
-    "pallet-timestamp/std",
-    "pallet-transaction-payment-rpc-runtime-api/std",
-    "pallet-transaction-payment/std",
-    "xcm/std",
 ]
-runtime-benchmarks = [
-    "frame-benchmarking/runtime-benchmarks",
-    "frame-support/runtime-benchmarks",
-    "frame-system-benchmarking/runtime-benchmarks",
-    "frame-system/runtime-benchmarks",
-    "pallet-assets/runtime-benchmarks",
-    "pallet-balances/runtime-benchmarks",
-    "pallet-grandpa/runtime-benchmarks",
-    "pallet-sudo/runtime-benchmarks",
-    "pallet-hybrid-orderbook/runtime-benchmarks",
-    "pallet-timestamp/runtime-benchmarks",
-    "sp-runtime/runtime-benchmarks",
-]
-try-runtime = [
-    "frame-try-runtime/try-runtime",
-    "frame-executive/try-runtime",
-    "frame-system/try-runtime",
-    "frame-support/try-runtime",
-    "pallet-assets/try-runtime",
-    "pallet-aura/try-runtime",
-    "pallet-balances/try-runtime",
-    "pallet-grandpa/try-runtime",
-    "pallet-sudo/try-runtime",
-    "pallet-hybrid-orderbook/try-runtime",
-    "pallet-timestamp/try-runtime",
-    "pallet-transaction-payment/try-runtime",
-]
-experimental = [
-    "frame-support/experimental",
-    "frame-system/experimental",
-    "pallet-example-tasks/experimental",
-]
+runtime-benchmarks = ["polkadot-sdk/runtime-benchmarks"]
+try-runtime = ["polkadot-sdk/try-runtime"]
